pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "analytics-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.8",
    "notebook>=7.5.3",
    "pandas>=2.3.3",
    "pyarrow>=23.0.1",
    "pyyaml>=6.0.3",
    "rank-bm25>=0.2.2",
    "tabulate>=0.9.0",
]

[tool.setuptools]
packages = ["agent"]
